refactor: extract new type to separate file
