+++ src/index.ts
@@ -5,6 +5,7 @@ export * from './types'
 export * from './utils'
 export * from './cache'
 export * from './rules/jsonschema.const'
+export * from './rules/jsonschema.types'
 export * from './rules/openapi.const'
 export * from './rules/asyncapi.const'
 export * from './rules/graphapi.const'
