+++ src/rules/jsonschema.types.ts
@@ -1,0 +1,12 @@
+import {
+  JSON_SCHEMA_PROPERTY_EXCLUSIVE_MAXIMUM,
+  JSON_SCHEMA_PROPERTY_EXCLUSIVE_MINIMUM,
+  JSON_SCHEMA_PROPERTY_MAXIMUM,
+  JSON_SCHEMA_PROPERTY_MINIMUM
+} from "./jsonschema.const";
+
+export type JsonSchemaNumericValidationKeywordsType =
+  typeof JSON_SCHEMA_PROPERTY_MINIMUM
+  | typeof JSON_SCHEMA_PROPERTY_MAXIMUM
+  | typeof JSON_SCHEMA_PROPERTY_EXCLUSIVE_MINIMUM
+  | typeof JSON_SCHEMA_PROPERTY_EXCLUSIVE_MAXIMUM
