+++ src/unifies/redundant-numeric-bounds.ts
@@ -5,8 +5,10 @@ import {
   JSON_SCHEMA_PROPERTY_EXCLUSIVE_MINIMUM,
   JSON_SCHEMA_PROPERTY_MAXIMUM,
   JSON_SCHEMA_PROPERTY_MINIMUM,
-  JsonSchemaNumericValidationKeywordsType,
 } from '../rules/jsonschema.const'
+import {
+  JsonSchemaNumericValidationKeywordsType,
+} from '../rules/jsonschema.types'
 import { cleanOrigins } from '../origins'
 import { isBroken, isPureCombiner } from './type'
 import { isNumber } from '../utils'
